+++ docs/NEWS.md
@@ -8,6 +8,7 @@
  * fix: prefix completion time ("`--fineta`") with *FIN* rather than *ETA* ([#43](https://codeberg.org/a-j-wood/pv/issues/43))
  * fix: surround average rate ("`--average-rate`") with brackets rather than square brackets
  * fix: correct a memory leak in "`--watchfd PID`"
+ * fix: make "`--direct-io`" work correctly with "`--output`" instead of assuming stdout
  * fix: write UTC timestamps in debugging mode to avoid lockups in signal handlers
  * cleanup: removed TODO.md, since it's just an outdated copy of the issue tracker
  * cleanup: re-ordered structure members to reduce padding